requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
Pillow==10.0.1
Flask-Mail==0.9.1
Flask-Compress==1.14
pycryptodome>=3.23.0
python-dotenv==1.0.0

# Testing (optional)
pytest>=7.0.0
pytest-flask>=1.3.0
pytest-cov>=4.0.0

# Production Server (choose one)
gunicorn>=21.0.0
# waitress>=2.1.0  # Alternative for Windows
# uwsgi>=2.0.0  # Alternative for Linux